centralized error handling add

+++ backend/controllers/cards.js
@@ -1,75 +1,66 @@
 const Card = require('../models/card');
-const errorHandler = require('../errors/errorHandler');
+const NotFoundError = require('../errors/notFoundError');
+const ForbiddenError = require('../errors/ForbiddenError');
+const BadRequestError = require('../errors/BadRequestError');
 
-module.exports.getCards = (req, res) => {
+module.exports.getCards = (req, res, next) => {
   Card.find({})
     .then((cards) => res.send(cards))
-    .catch((error) => errorHandler(error, res));
+    .catch(next);
 };
 
-module.exports.createCard = (req, res) => {
+module.exports.createCard = (req, res, next) => {
   const { name, link } = req.body;
   Card.create({ name, link, owner: req.user })
     .then((card) => res.send(card))
-    .catch((error) => errorHandler(error, res));
+    .catch((err) => {
+      if(err.name === 'ValidationError') {
+        next(new BadRequestError(err.message));
+      };
+
+      next(err);
+    });
 };
 
-module.exports.deleteCard = (req, res) => {
-  // Card.findById(req.params.cardId)
-  //   .orFail(() => {
-  //     const error = new Error('Card not found');
-  //     error.statusCode = 404;
-  //     throw error;
-  //   })
-  //   .then((card) => {
-  //     if (card.owner !== req.user) {
-  //       const error = new Error('You dont have rights to delete this card');
-  //       error.statusCode = 403;
-  //       throw error;
-  //     } else {
-  //       Card.findByIdAndRemove(req.params.cardId)
-  //         .then((card) => res.send(card))
-  //         .catch((error) => errorHandler(error, res));
-  //       }
-  //   })
-  //   .catch((error) => errorHandler(error, res));
-
-  Card.findByIdAndRemove(req.params.cardId)
+module.exports.deleteCard = (req, res, next) => {
+  Card.findById(req.params.cardId)
     .orFail(() => {
-      const error = new Error('Card not found');
-      error.statusCode = 404;
-      throw error;
+      throw new NotFoundError('Card not found');
     })
-    .then((card) => res.send(card))
-    .catch((error) => errorHandler(error, res));
+    .then((card) => {
+      if (card.owner != req.user) {
+        throw new ForbiddenError('You dont have rights to delete this card');
+      } else {
+        Card.findByIdAndRemove(req.params.cardId)
+          .then((card) => res.send(card))
+          .catch(next);
+        }
+    })
+    .catch(next);
 };
 
-module.exports.likeCard = (req, res) => {
+module.exports.likeCard = (req, res, next) => {
   Card.findByIdAndUpdate(
     req.params.cardId,
     { $addToSet: { likes: req.user } },
     { new: true },
   )
     .orFail(() => {
-      const error = new Error('Card not found');
-      error.statusCode = 404;
-      throw error;
+      throw new NotFoundError('Card not found');
     })
     .then((card) => res.send(card))
-    .catch((error) => errorHandler(error, res));
+    .catch(next);
 };
 
-module.exports.dislikeCard = (req, res) => {
+module.exports.dislikeCard = (req, res, next) => {
   Card.findByIdAndUpdate(
     req.params.cardId,
     { $pull: { likes: req.user } },
     { new: true },
   )
     .orFail(() => {
-      const error = new Error('Card not found');
-      error.statusCode = 404;
-      throw error;
+      throw new NotFoundError('Card not found');
     })
     .then((card) => res.send(card))
-    .catch((error) => errorHandler(error, res));
+    .catch(next);
 };
